refactor(dag-cli,agent-framework): SEC-006 split studio server and document the new exports
The SEC-006 request guards pushed studio/http-server.ts past its frozen size baseline, and the scan
is right that the answer is to split rather than extend. Extracted request-guards.ts (Host check +
canonical path containment), http-io.ts (readBody/jsonReply/sendSSE) and provider-routes.ts;
http-server.ts is now 458 lines against a 515 baseline. Behaviour unchanged — the same 31 studio
tests pass.

Also documents isSafeSessionId/assertSafeSessionId in the agent-session and agent-framework SPEC
Public API tables, and folds the framework re-export onto the existing InteractiveSession line so
src/index.ts stays within its baseline.
