+++ packages/agent-framework/docs/SPEC.md
@@ -151,6 +151,8 @@ Core classes and functions exported from `@robota-sdk/agent-framework`:
 | Export                                      | Kind     | Description                                                                                                                                                                                                                                                                          |
 | ------------------------------------------- | -------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
 | `InteractiveSession`                        | class    | Primary SDK entry point; event-driven session wrapper                                                                                                                                                                                                                                |
+| `isSafeSessionId`                           | function | SEC-006: whether a session id is safe as a filesystem path component; re-exported from `agent-session` so a consumer holding an id from an untrusted source can reject it without skipping a layer                                                                                   |
+| `assertSafeSessionId`                       | function | SEC-006: throwing form of `isSafeSessionId`                                                                                                                                                                                                                                          |
 | `resolveRoleModel`                          | function | SELFHOST-006: resolve a role's PRIMARY `IModelRef` from a `TRoleModelMap` (opaque key; undefined if unmapped)                                                                                                                                                                        |
 | `resolveRoleFallbackChain`                  | function | SELFHOST-006: the role's full ordered fallback chain (primary first); empty if unmapped                                                                                                                                                                                              |
 | `runWithRoleFallback`                       | function | SELFHOST-006: walk a role's fallback chain, trying each `IModelRef` over the provider DIP until one succeeds (alternate provider+model on error)                                                                                                                                     |
