+++ packages/agent-framework/src/index.ts
@@ -2,8 +2,7 @@
 // Provider-neutral. InteractiveSession is the single entry point.
 
 // ── InteractiveSession (primary API) ────────────────────────
-export { InteractiveSession } from './interactive/index.js';
-export { assertSafeSessionId, isSafeSessionId } from './interactive/index.js';
+export { InteractiveSession, assertSafeSessionId, isSafeSessionId } from './interactive/index.js';
 
 // ── Autonomous goal pursuit (GOAL-001) ──────────────────────
 export {
