+++ packages/dag-cli/src/studio/request-guards.ts
@@ -1,0 +1,73 @@
+/**
+ * SEC-006 — request guards for the local `dag studio` HTTP server.
+ *
+ * The studio binds 127.0.0.1, which is NOT an access control: any web page the developer visits can
+ * issue cross-origin requests to it, and before this module `/api/dag`, `/api/run` and `/api/validate`
+ * each resolved a client-supplied `file` against `cwd` with no containment — so a visited page could
+ * read arbitrary JSON from the machine and, via `/api/run`, make the server EXECUTE an arbitrary DAG.
+ *
+ * Kept in its own module so the containment and origin checks are reviewable on their own, rather than
+ * buried among the route handlers.
+ */
+import { realpathSync } from 'node:fs';
+import { basename, dirname, join, resolve, sep } from 'node:path';
+
+const LOOPBACK_HOSTS = new Set(['localhost', '127.0.0.1', '::1', '[::1]']);
+
+/**
+ * Whether the request `Host` header is a loopback name (port stripped) — closes DNS rebinding.
+ * A missing `Host` is rejected (a well-formed HTTP/1.1 client always sends one).
+ */
+export function isLoopbackHostHeader(host: string | undefined): boolean {
+  if (!host) return false;
+  const name = host.startsWith('[')
+    ? host.slice(0, host.indexOf(']') + 1)
+    : host.slice(0, host.lastIndexOf(':') === -1 ? host.length : host.lastIndexOf(':'));
+  return LOOPBACK_HOSTS.has(name);
+}
+
+/**
+ * Canonicalize `p` so symlinks are resolved, tolerating a path that does not exist yet.
+ *
+ * `realpathSync` throws on a missing path, but a client may legitimately reference a file that is not
+ * there (the caller then reports a read error). So we walk up to the deepest ancestor that DOES exist,
+ * canonicalize that, and re-attach the remaining segments — the re-attached tail cannot itself be a
+ * symlink, since it does not exist.
+ */
+function canonicalize(p: string): string {
+  const tail: string[] = [];
+  let current = p;
+  for (;;) {
+    try {
+      return join(realpathSync(current), ...tail);
+    } catch {
+      const parent = dirname(current);
+      // allow-fallback: nothing along the chain exists (or is readable) — fall back to the lexical
+      // path, which is still fully normalized by `resolve()` and therefore safe to compare.
+      if (parent === current) return p;
+      tail.unshift(basename(current));
+      current = parent;
+    }
+  }
+}
+
+/**
+ * Resolve a CLIENT-SUPPLIED `file` against `cwd` and refuse anything that escapes it.
+ *
+ * Containment is decided on the CANONICAL (symlink-resolved) paths: `resolve()` alone is purely
+ * lexical, so `<cwd>/link/x.dag.json` where `link -> /etc` passed a `startsWith` check while the
+ * subsequent `readFile` followed the link out of the working directory. Both sides are canonicalized so
+ * a cwd that is itself behind a symlink (macOS `/tmp` -> `/private/tmp`) still matches.
+ */
+export function resolveContainedFile(
+  file: string,
+  cwd: string,
+): { ok: true; path: string } | { ok: false; message: string } {
+  const resolved = resolve(cwd, file);
+  const canonical = canonicalize(resolved);
+  const canonicalCwd = canonicalize(resolve(cwd));
+  if (canonical !== canonicalCwd && !canonical.startsWith(canonicalCwd + sep)) {
+    return { ok: false, message: `Access denied: "${file}" is outside the working directory` };
+  }
+  return { ok: true, path: resolved };
+}
